Fix Linux build blocker, ID-truncation lookup bug, and double-.json save bug found in scheduled code scan #15 (#17)
- WorldEditor.h included <Helpers.h> while the file on disk is
  helpers.h; case-sensitive filesystems (Linux/macOS) failed to
  compile. Confirmed by a clean CMake configure/build in this
  environment before the fix, and a full compile pass afterward.
- FindGameObjectByID/FindEntityByID/FindInteractableByID (in both
  GameMap and the WorldEditor free-function duplicates) took the id
  as uint8_t, truncating the uint64_t object/entity/interactable ids.
  Any id above 255 would silently fail to look up or alias onto the
  wrong object — this fed straight into WorldEditor::getSelectedObject
  and the interactable activator lookup. Restored uint64_t.
- SaveGame/LoadGame append ".json" to the filename internally, but
  mainMenu.cpp and gameMain.cpp passed "mainMenu.json"/"backup.json",
  producing "mainMenu.json.json" / "backup.json.json" — the exact
  issue flagged (but left unresolved) in the "Find World Files"
  commit. Main menu's scene load and the on-close backup save were
  both silently failing/writing to the wrong file. Passed the bare
  name instead.

+++ src/gameLayer/gameMap.h
@@ -26,9 +26,9 @@ struct GameMap {
 	void removeInteractable(InteractableObject* object);
 
 	// Find Data
-	GameObject* FindGameObjectByID(std::uint8_t id);
-	Entity* FindEntityByID(std::uint8_t id);
-	InteractableObject* FindInteractableByID(std::uint8_t id);
+	GameObject* FindGameObjectByID(std::uint64_t id);
+	Entity* FindEntityByID(std::uint64_t id);
+	InteractableObject* FindInteractableByID(std::uint64_t id);
 
 	// Return Data
 	Vector3 getMapSize() const { return Vector3(size.x, size.y, size.z);}
